refactor : (MVC) 입력과 그 핵심기능을 클래스로 분리
