+++ src/main/java/lotto/view/WinningInput.java
@@ -3,18 +3,15 @@ package lotto.view;
 import static camp.nextstep.edu.missionutils.Console.readLine;
 
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Stream;
 
 import lotto.domain.ErrorMessage;
+import lotto.domain.Message;
 
 public class WinningInput {
     public static String winningInput;
     public static List<Integer> winningNumbers;
-    public static int winningCount;
     private static int[] winningIntArr;
 
     public static void inputWinningNumber() {
@@ -24,15 +21,6 @@ public class WinningInput {
         System.out.println();
         winningNumberToInt(winningInput);
         winningIntArrToList(winningIntArr);
-    }
-
-    public static void getWinningResult(List<List<Integer>> lottos, List<Integer> winningNumbers) {
-        Map<List<Integer>, Integer> map = new HashMap<>();
-        for (List<Integer> check : lottos) {
-            winningCount = 0;
-            checkCount(check, winningNumbers, map);
-        }
-        winningCount = Collections.max(map.values());
     }
 
     private static void winningNumberToInt(String winningInput) {
@@ -47,15 +35,6 @@ public class WinningInput {
         winningNumbers = new ArrayList<>(winningIntArr.length);
         for (int i : winningIntArr) {
             winningNumbers.add(i);
-        }
-    }
-
-    private static void checkCount(List<Integer> check, List<Integer> winningNumbers, Map<List<Integer>, Integer> map) {
-        for (int i = 0; i < check.size(); i++) {
-            if (check.contains(winningNumbers.get(i))) {
-                winningCount++;
-                map.put(check, winningCount);
-            }
         }
     }
